slug, timestampable and data validation

+++ app/src/Controller/CategoryController.php
@@ -161,6 +161,14 @@ class CategoryController extends AbstractController
     )]
     public function delete(Request $request, Category $category): Response
     {
+        if(!$this->categoryService->canBeDeleted($category)) {
+            $this->addFlash(
+                'warning',
+                $this->translator->trans('message.category_contains_tasks')
+            );
+
+            return $this->redirectToRoute('category_index');
+        }
         $form = $this->createForm(FormType::class, $category, [
             'method' => 'DELETE',
             'action' => $this->generateUrl('category_delete', ['id' => $category->getId()]),
